requirements.txt
# Core pipeline
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.11
pandas>=2.0.0
numpy>=1.26.0

# Embeddings + retrieval
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0           # use faiss-gpu if you have CUDA

# LLM
anthropic>=0.28.0

# UI
streamlit>=1.35.0
